Add two compile time parameters SHEAR_CORRECTION and HEIGHT_CORRECTION.
If the image is sheared, which seems to appear on widescreen monitors
were the width is not a multiple of the width of a character, you can
add SHEAR_CORRECTION=? where ? as an integer. For example, I use
SHEAR_CORRECTION=10 on my laptop.

If the height of the image does not fill the height of the screen,
which seems to appear on widescreen monitors were the height is not
a multiple of the height of a character, you can add HEIGHT_CORRECTION=?
where ? is the number of missing pixels. For example, I use
HEIGHT_CORRECTION=8 on my netbook.

I have not found any automatic way to fix this, but I hope you can.

Signed-off-by: Mattias Andrée <maandree@operamail.com>